+++ docs/changelog.rst
@@ -1,6 +1,9 @@
 Changelog
 =========
 
+Version 0.4
+-----------
+
 Version 0.3
 -----------